+++ docs/src/components/landing-page/DemoVideo.js
@@ -31,7 +31,7 @@ export default function DemoVideo() {
       <img
         src={WorldDotPng}
         style={{ marginTop: '-50rem' }}
-        className="absolute top-0 right-0"
+        className="absolute top-0 right-0 lg:block hidden"
       />
       <div className="m-auto flex flex-col z-10">
         <h3 className="m-auto text-white text-xl -mt-y-1/4">
@@ -42,7 +42,7 @@ export default function DemoVideo() {
         </h3>
         <div className="flex flex-wrap justify-around gap-8 mt-8">
           <iframe
-            className="my-auto w-96 h-72"
+            className="my-auto md:w-96 md:h-72 w-56 h-32"
             src="https://www.youtube-nocookie.com/embed/o4jrNeNeFmM"
             title="YouTube video player"
             frameBorder="0"
@@ -50,19 +50,19 @@ export default function DemoVideo() {
             allowFullScreen={true}
           />
           <div className="m-auto flex flex-col max-w-sm">
-            <p className="text-white text-4xl leading-none">
+            <p className="text-white text-4xl leading-none lg:text-left text-center">
               Sleep well knowing that your{' '}
               <span className="bg-gradient-to-r from-purple-monika to-aqua-monika bg-clip-text text-transparent font-bold">
                 server is working as it should be
               </span>
             </p>
-            <p className="text-white mt-4 leading-tight">
+            <p className="text-white mt-4 leading-tight lg:text-left text-center">
               Monitor every part of your web app using a simple JSON
               configuration file. Get alert not only when your site is down, but
               also when it's slow.
             </p>
             <ButtonLink
-              className="mt-6 w-40 leading-snug font-semibold text-sm text-center"
+              className="mt-6 w-40 lg:ml-0 mx-auto leading-snug font-semibold text-sm text-center"
               href="/quick-start"
               rel="noopener noreferrer"
             >
